+++ apt/apt-pkg/rpm/rpmlistparser.h
@@ -67,9 +67,9 @@ class rpmListParser : public pkgCacheGenerator::ListParser
    virtual unsigned short VersionHash() override;
    virtual bool UsePackage(pkgCache::PkgIterator &Pkg,
 			   pkgCache::VerIterator &Ver) override;
-   virtual unsigned long long Offset() override
+   virtual unsigned long Offset() override
 	{return Handler->Offset();};
-   virtual unsigned long long Size() override;
+   virtual unsigned long Size() override;
    virtual unsigned long Flags() override;
 
    virtual bool OrderedOffset() override
